ca.c
This problem is based on Round Robin algorithm. Basically, in this problem we have been provide with or we have to consider the arrival time and burst time requirement of the process. In which scheduler is scheduling the process by interrupting the process after every 6 unit of time, which is also called as the time quantum for each process and we need to consider the completion of the process in this iteration. After that the scheduler then check for the number of process waiting for the process and the processor to the process but interrupting the processor every 10 unit of time and we need to consider the completion of the process in this iteration. The scheduler checks the number of processes waiting in the queue for the processor after the second iteration and gives the processor to the process which needs more time to complete than the other processes to go in the terminated state. Now we have to develop a scenario in which processes will be submitted to the processor and also have to compute the scheduler performance by calculating the waiting time for the processes, turnaround time for the processes and average waiting and average turnaround time.